Update README.md
This is the manifesto of my crazy endeavor. May it help me to find joy in the success that I already know, and never to lose it regardless of the outcome of my coding endeavor. May I write programs with the intent to LIVE WELL and help others, not to climb the ranks or collect accolades.